Create endpoint Update Testimonials
