+++ src/controllers/testimonial-controller.js
@@ -34,7 +34,25 @@ const createNewTestimony = async (req, res) => {
     }
 }
 
+const updateTestimony = async (req,res,next) => {
+    try {
+        const { id } = req.params;
+        const { newContent } = req.body;  //TODO Destructurar el body
+        const testimonyUpdated = await testimonialService.updateTestimony(id, newContent);
+        if (testimonyUpdated) {
+            throw new Error('El testimonio que desea modificar no existe.');
+        }
+        res.status(200).json(testimonyUpdated);
+    }
+    catch (err) {
+        console.log(err)
+        res.status(500).json(err)
+    }
+}
+
+
 module.exports = {
     getAllTestimonials,
-    createNewTestimony
+    createNewTestimony,
+    updateTestimony
 }
