+++ src/services/testimonial-service.js
@@ -18,8 +18,17 @@ const createTestimony = async( testimony ) =>{
     return testimonyStored;
   }
 
+  const updateTestimony = async (id, newContent) => {
+    const testimonyUpdated = await testimonialRepository.updateTestimony(id, newContent);
+    if (testimonyUpdated) {
+        throw new Error('El testimonio que desea modificar no existe.');
+    }
+    return testimonyUpdated;
+}
+
 
 module.exports= {
     getAllTestimonials,
-    createTestimony
+    createTestimony,
+    updateTestimony
 }
